% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DHT11_writeMemory}
\alias{DHT11_writeMemory}
\title{Write Data from DHT11 Sensor into Shared Memory}
\usage{
DHT11_writeMemory(timeDelay = 1000L, pin = 0L)
}
\arguments{
\item{v1}{First value}

\item{v2}{Second value}
}
\value{
Product of v1 and v2
}
\description{
Write Data from DHT11 Sensor into Shared Memory
}
