% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LL_digitalWrite}
\alias{LL_digitalWrite}
\title{Dummy Docs}
\usage{
LL_digitalWrite(pin, mode)
}
\arguments{
\item{pin}{An integer vector}

\item{mode}{A logical vector}
}
\description{
This function returns XYZ. 
BLAHBLAH
}
