% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LL_pinMode}
\alias{LL_pinMode}
\title{Dummy Docs}
\usage{
LL_pinMode(pin, mode)
}
\arguments{
\item{pin}{An integer vector}

\item{mode}{A string vector}
}
\description{
This function returns XYZ. 
BLAHBLAH
}
