% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LL_pullUpDnControl}
\alias{LL_pullUpDnControl}
\title{Dummy Docs}
\usage{
LL_pullUpDnControl(pin, pud)
}
\arguments{
\item{pin}{An integer vector}

\item{pud}{A string vector}
}
\description{
This function returns XYZ. 
BLAHBLAH
}
