% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PhotoRes_readMemory}
\alias{PhotoRes_readMemory}
\title{Read Data from Shared Memory for Photo Resistor Sensor}
\usage{
PhotoRes_readMemory(n = 1L)
}
\arguments{
\item{n}{An integer vector, specifying number of data points to read from the shared memory.}
}
\description{
This function returns an integer vector.
}
