% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PhotoRes_writeMemory}
\alias{PhotoRes_writeMemory}
\title{Write Data into Memory for Photo-Resistor Sensor}
\usage{
PhotoRes_writeMemory(pin = 7L, length = 100L, timeDelay = 1000L)
}
\arguments{
\item{pin}{An integer vector, specifying the pin for connecting the DHT11 sensor.}

\item{length}{An integer vector, specifying the amount of memory}

\item{timeDelay}{An integer vector, specifying the number of milliseconds for delay.}
}
\description{
This function returns NULL. 
It is a separate instance 
everything else as FALSE.
}
