% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simpleCapture}
\alias{simpleCapture}
\title{Dummy Docs}
\usage{
simpleCapture(
  path = "/home/pi/Pictures/raspicam_image.ppm",
  width = 1L,
  height = 1L
)
}
\arguments{
\item{path}{An integer vector}

\item{width}{An integer vector}

\item{height}{An integer vector}
}
\description{
This function returns XYZ. 
BLAHBLAH
}
