% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testingMQ2}
\alias{testingMQ2}
\title{Dummy Docs}
\usage{
testingMQ2(
  pin = 0L,
  SPICLK = 14L,
  SPIMISO = 13L,
  SPIMOSI = 12L,
  SPICS = 10L,
  mq2_dpin = 25L,
  mq2_apin = 0L
)
}
\arguments{
\item{pin}{An integer vector}

\item{SPICLK}{An integer vector}

\item{SPIMISO}{An integer vector}

\item{SPIMOSI}{An integer vector}

\item{SPICS}{An integer vector}

\item{mq2_dpin}{An integer vector}

\item{mq2_apin}{An integer vector}
}
\description{
This function returns XYZ. 
BLAHBLAH
}
