% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DHT11_writeMemory}
\alias{DHT11_writeMemory}
\title{Write Data into Memory for DHT11 Sensor}
\usage{
DHT11_writeMemory(pin = 0L, length = 100L, timeDelay = 1000L)
}
\arguments{
\item{pin}{An integer vector, specifying the pin for connecting the DHT11 sensor.}

\item{length}{An integer vector, specifying the amount of memory}

\item{timeDelay}{An integer vector, specifying the number of milliseconds for delay.}
}
\description{
This function returns NULL. 
It is a separate instance 
everything else as FALSE.
}
